+++ pypit/flatfield.py
@@ -9,7 +9,7 @@ from pypit import msgs
 from pypit import processimages
 from pypit import masterframe
 from pypit.core import arprocimg
-from pypit import arproc
+from pypit.core import arflat
 
 from pypit import ardebug as debugger
 
@@ -55,6 +55,9 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
         # Attributes (set after init)
         self.frametype = frametype
 
+        # Key outputs
+        self.mspixelflat = None
+
         # Settings
         # The copy allows up to update settings with user settings without changing the original
         if settings is None:
@@ -74,24 +77,34 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
         # MasterFrames
         masterframe.MasterFrame.__init__(self, self.frametype, self.setup, self.settings)
 
-    def run(self, datasec_img, lordloc, rordloc, pixwid, slitpix, tilts, settings_det, trim=True, armed=False):
-        #ntcky=settings.argflag['reduce']['flatfield']['params'][0])
+    def apply_gain(self, datasec_img, settings_det):
+        # Apply gain (instead of ampsec scale)
+        self.mspixelflat *= arprocimg.gain_frame(datasec_img, settings_det['numamplifiers'], settings_det['gain'])
+
+    def build_pixflat(self, trim=True):
         # Generate the image
         self.mspixelflat = self.process(bias_subtract=self.msbias, trim=trim)
 
-        # Apply gain (instead of ampsec scale)
-        self.mspixelflat *= arprocimg.gain_frame(datasec_img, settings_det['numamplifiers'], settings_det['gain'])
+    def run(self, datasec_img, lordloc, rordloc, pixwid, slitpix, tilts, settings_det, armed=False):
+        #ntcky=settings.argflag['reduce']['flatfield']['params'][0])
+
+        # Build the pixel flat (as needed)
+        if self.mspixelflat is None:
+            self.mspixelflat = self.build_pixflat()
+
+        # Apply gain
+        self.apply_gain(datasec_img, settings_det)
 
         # Normalize the flat field
         msgs.info("Normalizing the pixel flat")
         slit_profiles, mstracenrm, msblaze, flat_ext1d, extrap_slit = \
-            arproc.slit_profile(self.mspixelflat, datasec_img, lordloc, rordloc, pixwid, slitpix,
+            arflat.slit_profile(self.mspixelflat, datasec_img, lordloc, rordloc, pixwid, slitpix,
                                 self.det, tilts, ntcky=self.settings['flatfield']['params'][0])
 
         # If some slit profiles/blaze functions need to be extrapolated, do that now
         if armed:  #settings.spect['mosaic']['reduction'] == 'AMRED':
             if np.sum(extrap_slit) != 0.0:
-                slit_profiles, mstracenrm, msblaze = arproc.slit_profile_pca(
+                slit_profiles, mstracenrm, msblaze = arflat.slit_profile_pca(
                     self.mspixelflat, tilts, msblaze, extrap_slit, slit_profiles,
                     lordloc, rordloc, pixwid, slitpix)
